Structs
I have experimented with structs by creating a Rectangle type that has Length and Width fields.

I have also declared a function within it with the NEW keyword to define how to instantiate it.

I then created a function that takes in a Rectangle type and returns its area by multiplying its two fields, and used it.