Note for the Previous Commit
I have used a different approach than the one that the book prescribes.

I have used recursive function rather than a while loop to keep asking the user what he would like to do. The recursive function got quite large. The reason for this is that the initial solution I had in mind introduced a cyclical dependency, and to enable the function calls up and down the chain I have moved what was once tiny seperate functions into one large function. This is probably not the best way of dealing with the problem.

I have also used shadowing to pass the new value of the user's balance both into the printf function to display it back, and into the new account types that we are returning from both the withdraw and deposit function.