Lesson 15 Scratchpad
I have created a script file to use it to copy and paste the code illustrations in the book, take notes and experiment with them.

Near the end of Lesson 15, I have decided to move onto the Capstone of this Unit as I believe that I'll be able to deal with it at this point. If not, I will return back to the lessons to find a way through the solution.

Next on, the Unit 4 Capstone.