Capstone Done (Except Audit)
To implement the audit functionality, I will look into the book's exercise files repository.

Since I haven't been using the book's exercise files, and barely reading the instructions on the capstones apart from the 'identifying the problem' sections, my app started to look very different. Consequently, it was getting harder and harder to follow the instructions of the book.

I have abandoned my previous approach and adopted the way the book prescribes. A lot of functions have changed and now we're using the Seq module to keep the transaction look running until the user presses 'x' in their keyboard.  We're using yield in the last function of the pipeline to enable this.

One downside to doing this was that now we don't have only one impure function but most of the functions are impure. As I understand that this is not what's important to get accustomed to at this point in the book, that's not a problem at all.