Unit 4 Capstone Finished
I have finished the capstone for unit 4 by modeling my code after the corresponding code in the book's exercise files.

One difference is though, that I don't have seperate files containing blocks of code. I have one file containing all the code with modules that have AutoOpen attributes.

Now the app saves the user to the disk and reads it back again if the user input for the Customer Name corresponds to a Customer record that has already been saved.