Seperating Code into Seperate Files and Modules
I have edited the code of the application to be in seperate files inside their respective modules. Something very close to the example solution in the exercise files, except that I have used modules only instead of namespaces.

I have examined in what fashion the code is seperated and have an idea of how I might do something similar when a file get's too long. I also will be better able to follow the capstone if it is based on this one since now I have a solution that is closer to the one that the book prescribes.