Essential Functions Set in Place
I have created a console application and a Script.fsx file as a scratchpad.

In the script file, I have written some of the most fundamental functions that I will need to program the game, alongside the domain model that I'm going to need.

Some functions are: generateDeck, shuffleDeck, drawCard, and dealHand.