Domain Model, 'startGame' and 'generateDeck' functions.
In last commit I was in the beggining of creating the dealHands function. Instead I'm now creating a startGame function.

I had to make some changes to the model state. We have two records, both of which are only has one field cald Cards which is a list of cards, Hand and Deck. I have created wrappers to easily identify which is which.

I have altered to generateDeck function to actually return a Deck type rather than a list of cards to fit well with what I have in mind for the startGame function.

I am using Recursion and Options heavily as probably a quick solution to the problems. I will start to think about cleaner alternative methods once I have the game running on console.