Starting the Game
I have developed and tested all the functions that are necessary to start the game, utilizing heavily the domain model, recursive functions and option types.

Nothing is done yet about the actual play logic. We have functions that when used in combination and given the correct input returns a List of Players including the cards in their hands, and the rest of the cards in a shuffled state.

Some code here and there gets unnecessarily complicated in my opinion and I will refactor it as soon as I have a game running. However, nothing is so hard to read and I have used a lot of comments.