Active Patterns, tryServePlayer function
I have utilized a few Active Patterns, both to experiment with them and because the code was getting really hard to read with some of the pattern matches. I have only utilized them for the last function for now, but I will keep looking for blocks of code that are better off using them.

I have defined a tryServePlayer function that takes in a player and a desk. It returns a new player with its Cards appended by a card from the Deck and a new deck that has the same Card removed, if the deck has at least one Card. (Option type)